pyproject.toml
[tool.poetry]
name = "ecommerce-data-warehouse"
version = "0.1.0"
description = ""
authors = ["gabrielcamba <gabrielcamba@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
faker = "^35.0.0"
python-dateutil = "^2.9.0.post0"
unidecode = "^1.3.8"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
